+++ config/max-lines-baseline.txt
@@ -12,14 +12,9 @@ inline src/main/automations/external-manager.ts
 inline src/main/automations/hermes-cron-output.ts
 inline src/main/browser/agent-browser-bridge.ts
 inline src/main/browser/browser-cookie-import.ts
-inline src/main/browser/browser-guest-ui.ts
 inline src/main/browser/browser-manager.ts
-inline src/main/browser/browser-screencast-stream.ts
-inline src/main/browser/browser-session-registry.ts
 inline src/main/browser/cdp-bridge.ts
-inline src/main/browser/cdp-ws-proxy.ts
 inline src/main/browser/grab-guest-script.ts
-inline src/main/browser/snapshot-engine.ts
 inline src/main/claude-accounts/runtime-auth-service.ts
 inline src/main/claude-accounts/service.ts
 inline src/main/claude-usage/scanner.ts
@@ -57,7 +52,6 @@ inline src/main/gitlab/work-item-details.ts
 inline src/main/hermes/hook-service.ts
 inline src/main/hooks.ts
 inline src/main/index.ts
-inline src/main/ipc/browser.ts
 inline src/main/ipc/crash-reporting.ts
 inline src/main/ipc/filesystem-auth.ts
 inline src/main/ipc/filesystem-mutations.ts
