+++ src/main/browser/browser-guest-grab-shortcuts.ts
@@ -1,0 +1,84 @@
+import { keybindingMatchesAction, type KeybindingOverrides } from '../../shared/keybindings'
+import type { ResolveRenderer } from './browser-guest-renderer-target'
+
+// Why: a focused guest never surfaces Cmd/Ctrl+C to the renderer; forward only when it wouldn't do a normal copy (no editable focus, no selection).
+export function setupGrabShortcutForwarding(args: {
+  browserTabId: string
+  guest: Electron.WebContents
+  resolveRenderer: ResolveRenderer
+  hasActiveGrabOp: (browserTabId: string) => boolean
+  getKeybindings?: () => KeybindingOverrides | undefined
+}): () => void {
+  const { browserTabId, guest, resolveRenderer, hasActiveGrabOp, getKeybindings } = args
+  const handler = (event: Electron.Event, input: Electron.Input): void => {
+    if (input.type !== 'keyDown') {
+      return
+    }
+    const bareKey = input.key.toLowerCase()
+    if (
+      !input.meta &&
+      !input.control &&
+      !input.alt &&
+      !input.shift &&
+      (bareKey === 'c' || bareKey === 's') &&
+      hasActiveGrabOp(browserTabId)
+    ) {
+      const renderer = resolveRenderer(browserTabId)
+      if (!renderer) {
+        return
+      }
+      // Why: a focused guest swallows bare keys; during an active grab pick, plain C/S are Orca's copy/screenshot, not page typing.
+      event.preventDefault()
+      renderer.send('browser:grabActionShortcut', { browserPageId: browserTabId, key: bareKey })
+      return
+    }
+
+    if (
+      !keybindingMatchesAction('browser.grabElement', input, process.platform, getKeybindings?.())
+    ) {
+      return
+    }
+
+    void guest
+      .executeJavaScript(`(() => {
+        const active = document.activeElement
+        const tag = active?.tagName
+        const isEditable =
+          active instanceof HTMLInputElement ||
+          active instanceof HTMLTextAreaElement ||
+          active?.isContentEditable === true ||
+          tag === 'SELECT' ||
+          tag === 'IFRAME'
+        if (isEditable) {
+          return false
+        }
+        const selection = window.getSelection()
+        return Boolean(selection && selection.type === 'Range' && selection.toString().trim().length > 0)
+          ? false
+          : true
+      })()`)
+      .then((shouldToggle) => {
+        if (!shouldToggle) {
+          return
+        }
+        event.preventDefault()
+        const renderer = resolveRenderer(browserTabId)
+        if (!renderer) {
+          return
+        }
+        renderer.send('browser:grabModeToggle', browserTabId)
+      })
+      .catch(() => {
+        // Why: shortcut forwarding is best-effort — guest teardown or a transient executeJavaScript failure must not break normal copy.
+      })
+  }
+
+  guest.on('before-input-event', handler)
+  return () => {
+    try {
+      guest.off('before-input-event', handler)
+    } catch {
+      // Why: browser tabs can briefly outlive the guest webContents during teardown, so cleanup is best-effort.
+    }
+  }
+}
