+++ src/main/browser/browser-guest-renderer-target.ts
@@ -1,0 +1,18 @@
+import { webContents } from 'electron'
+
+export type ResolveRenderer = (browserTabId: string) => Electron.WebContents | null
+
+export function resolveRendererWebContents(
+  rendererWebContentsIdByTabId: ReadonlyMap<string, number>,
+  browserTabId: string
+): Electron.WebContents | null {
+  const rendererWcId = rendererWebContentsIdByTabId.get(browserTabId)
+  if (!rendererWcId) {
+    return null
+  }
+  const renderer = webContents.fromId(rendererWcId)
+  if (!renderer || renderer.isDestroyed()) {
+    return null
+  }
+  return renderer
+}
