+++ src/main/browser/browser-screencast-debugger-command.ts
@@ -1,0 +1,25 @@
+import type { WebContents } from 'electron'
+
+const DEBUGGER_COMMAND_TIMEOUT_MS = 8_000
+
+export async function sendDebuggerCommand(
+  dbg: WebContents['debugger'],
+  method: string,
+  params: Record<string, unknown> = {}
+): Promise<unknown> {
+  let timeout: ReturnType<typeof setTimeout> | null = null
+  try {
+    return await Promise.race([
+      Promise.resolve().then(() => dbg.sendCommand(method, params)),
+      new Promise<never>((_, reject) => {
+        timeout = setTimeout(() => {
+          reject(new Error(`Timed out while running ${method}.`))
+        }, DEBUGGER_COMMAND_TIMEOUT_MS)
+      })
+    ])
+  } finally {
+    if (timeout) {
+      clearTimeout(timeout)
+    }
+  }
+}
