+++ src/main/ipc/browser-renderer-trust.ts
@@ -1,0 +1,25 @@
+let trustedBrowserRendererWebContentsId: number | null = null
+
+export function setTrustedBrowserRendererWebContentsId(webContentsId: number | null): void {
+  trustedBrowserRendererWebContentsId = webContentsId
+}
+
+export function isTrustedBrowserRenderer(sender: Electron.WebContents): boolean {
+  if (sender.isDestroyed() || sender.getType() !== 'window') {
+    return false
+  }
+  if (trustedBrowserRendererWebContentsId != null) {
+    return sender.id === trustedBrowserRendererWebContentsId
+  }
+
+  const senderUrl = sender.getURL()
+  if (process.env.ELECTRON_RENDERER_URL) {
+    try {
+      return new URL(senderUrl).origin === new URL(process.env.ELECTRON_RENDERER_URL).origin
+    } catch {
+      return false
+    }
+  }
+
+  return senderUrl.startsWith('file://')
+}
